+++ interpreter/src/builtins/constants.rs
@@ -1,4 +1,4 @@
 use core::Pallet;
 
-pub static PI: Pallet = Pallet::Float(std::f32::consts::PI);
-pub static E: Pallet = Pallet::Float(std::f32::consts::E);
+pub static PI: Pallet = Pallet::Float(std::f64::consts::PI);
+pub static E: Pallet = Pallet::Float(std::f64::consts::E);
